Make search augmentation resilient to libraries missing from the index
