+++ kifi-backend/modules/search/app/com/keepit/controllers/website/WebsiteSearchController.scala
@@ -89,14 +89,8 @@ class WebsiteSearchController @Inject() (
         }
 
         getAugmentedItems(augmentationCommander)(userId, kifiPlainResult).flatMap { augmentedItems =>
-          val (allSecondaryFields, userIds, libraryIds) = writesAugmentationFields(userId, maxKeepersShown, maxLibrariesShown, maxTagsShown, augmentedItems)
           val librarySearcher = libraryIndexer.getSearcher
-          val libraryRecordsAndVisibilityById = getLibraryRecordsAndVisibility(librarySearcher, libraryIds.toSet)
-
-          val futureUsers = {
-            val libraryOwnerIds = libraryRecordsAndVisibilityById.values.map(_._1.ownerId)
-            shoeboxClient.getBasicUsers(userIds ++ libraryOwnerIds)
-          }
+          val (allSecondaryFields, futureBasicUsersAndLibraries) = writesAugmentationFields(librarySearcher, userId, maxKeepersShown, maxLibrariesShown, maxTagsShown, augmentedItems)
 
           val futureJsHits = for {
             summaries <- futureUriSummaries
@@ -119,15 +113,9 @@ class WebsiteSearchController @Inject() (
             }
           }
           for {
-            usersById <- futureUsers
+            (users, libraries) <- futureBasicUsersAndLibraries
             jsHits <- futureJsHits
           } yield {
-            val libraries = libraryIds.map { libId =>
-              val (library, visibility) = libraryRecordsAndVisibilityById(libId)
-              val owner = usersById(library.ownerId)
-              makeBasicLibrary(library, visibility, owner)
-            }
-            val users = userIds.map(usersById(_))
             (jsHits, users, libraries)
           }
         }
